Fix some problemes with new system of annotations
