+++ config/routes/annotations.yaml
@@ -8,6 +8,11 @@ front_controllers:
     type: annotation
     prefix: '/'
 
+generator_controller:
+    resource: ../src/Controller/Generator
+    type: annotation
+    prefix: '/'
+
 back_controllers:
     resource: ../../src/Controller/Back
     type: annotation
