+++ src/Controller/Front/RegistrationController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Controller;
+namespace App\Controller\Front;
 
 use App\Service\EmailService;
 use App\Service\UserService;
